backend/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
openai-agents>=0.0.7
mcp[cli]>=1.2.0
